Apply the LFS b12.2 backspace patch

+++ projects/kbd-2.6.4/src/libkeymap/kernel.c
@@ -36,7 +36,7 @@ int lk_kernel_keys(struct lk_ctx *ctx, int fd)
 			ke.kb_index = (unsigned char) i;
 			ke.kb_value = 0;
 
-			if (ioctl(fd, KDGKBENT, (unsigned long)&ke)) {
+			if (ioctl(fd, KDGKBENT, &ke)) {
 				ERR(ctx, _("KDGKBENT: %s: error at index %d in table %d"),
 				    strerror(errno), i, t);
 				return -1;
@@ -68,7 +68,7 @@ int lk_kernel_funcs(struct lk_ctx *ctx, int fd)
 		}
 		kbs.kb_func = (unsigned char) i;
 
-		if (ioctl(fd, KDGKBSENT, (unsigned long)&kbs)) {
+		if (ioctl(fd, KDGKBSENT, &kbs)) {
 			ERR(ctx, _("KDGKBSENT: %s: Unable to get function key string"),
 			    strerror(errno));
 			return -1;
@@ -98,7 +98,7 @@ int lk_kernel_diacrs(struct lk_ctx *ctx, int fd)
 	int i;
 	struct lk_kbdiacr dcr;
 
-	if (ioctl(fd, request, (unsigned long)&kd)) {
+	if (ioctl(fd, request, &kd)) {
 		ERR(ctx, _("KDGKBDIACR(UC): %s: Unable to get accent table"),
 		    strerror(errno));
 		return -1;
